+++ otherlibs/site/src/dune_site_data.mli
@@ -1,0 +1,5 @@
+(** {2 encoded} *)
+
+val hardcoded_ocamlpath : string
+
+val stdlib_dir : string