+++ src/dune_rules/dune_package.ml
@@ -281,6 +281,8 @@ type t =
   { name : Package.Name.t
   ; entries : Entry.t Lib_name.Map.t
   ; version : string option
+  ; sections : Path.t Section.Map.t
+  ; sites : Section.t Section.Site.Map.t
   ; dir : Path.t
   }
 
@@ -288,6 +290,12 @@ let decode ~lang ~dir =
   let open Dune_lang.Decoder in
   let+ name = field "name" Package.Name.decode
   and+ version = field_o "version" string
+  and+ sections =
+    field ~default:[] "sections"
+      (repeat (pair (located Section.decode) Dpath.decode))
+  and+ sites =
+    field ~default:[] "sites"
+      (repeat (pair (located Section.Site.decode) Section.decode))
   and+ entries = leftover_fields_as_sums (Entry.cstrs ~lang ~dir) in
   let entries =
     List.map entries ~f:(fun e ->
@@ -308,7 +316,20 @@ let decode ~lang ~dir =
             (Lib_name.to_string name)
         ]
   in
-  { name; version; entries; dir }
+  let section_map of_list_map to_string sections =
+    match of_list_map sections ~f:(fun ((_, k), v) -> (k, v)) with
+    | Ok x -> x
+    | Error (s, ((loc, _), _), _) ->
+      User_error.raise ~loc
+        [ Pp.textf "The section %s appears multiple times" (to_string s) ]
+  in
+  let sections =
+    section_map Section.Map.of_list_map Section.to_string sections
+  in
+  let sites =
+    section_map Section.Site.Map.of_list_map Section.Site.to_string sites
+  in
+  { name; version; entries; dir; sections; sites }
 
 let () = Vfile.Lang.register Stanza.syntax ()
 
@@ -323,20 +344,21 @@ let prepend_version ~dune_version sexps =
   ]
   @ sexps
 
-let encode ~dune_version { entries; name; version; dir } =
-  let list s = Dune_lang.List s in
-  let sexp = [ list [ Dune_lang.atom "name"; Package.Name.encode name ] ] in
-  let sexp =
-    match version with
-    | None -> sexp
-    | Some version ->
-      sexp
-      @ [ List
-            [ Dune_lang.atom "version"
-            ; Dune_lang.atom_or_quoted_string version
-            ]
-        ]
+let encode ~dune_version { entries; name; version; dir; sections; sites } =
+  let open Dune_lang.Encoder in
+  let sections =
+    Section.Map.to_list (Section.Map.map ~f:Path.to_absolute_filename sections)
   in
+  let sites = Section.Site.Map.to_list sites in
+  let sexp =
+    record_fields
+      [ field "name" Package.Name.encode name
+      ; field_o "version" string version
+      ; field_l "sections" (pair Section.encode string) sections
+      ; field_l "sites" (pair Section.Site.encode Section.encode) sites
+      ]
+  in
+  let list s = Dune_lang.List s in
   let entries =
     Lib_name.Map.to_list entries
     |> List.map ~f:(fun (_name, e) ->
@@ -353,7 +375,7 @@ let encode ~dune_version { entries; name; version; dir } =
   in
   prepend_version ~dune_version (List.concat [ sexp; entries ])
 
-let to_dyn { entries; name; version; dir } =
+let to_dyn { entries; name; version; dir; sections; sites } =
   let open Dyn.Encoder in
   record
     [ ( "entries"
@@ -361,6 +383,8 @@ let to_dyn { entries; name; version; dir } =
     ; ("name", Package.Name.to_dyn name)
     ; ("version", option string version)
     ; ("dir", Path.to_dyn dir)
+    ; ("sections", Section.Map.to_dyn Path.to_dyn sections)
+    ; ("sites", Section.Site.Map.to_dyn Section.to_dyn sites)
     ]
 
 module Or_meta = struct