+++ src/dune_rules/findlib/findlib.ml
@@ -164,6 +164,8 @@ let builtin_for_dune : Dune_package.t =
   ; entries = Lib_name.Map.singleton (Dune_package.Entry.name entry) entry
   ; version = None
   ; dir = Path.root
+  ; sections = Section.Map.empty
+  ; sites = Section.Site.Map.empty
   }
 
 type db =
@@ -174,6 +176,8 @@ type db =
       (Package.Name.t, (Dune_package.t, Unavailable_reason.t) result) Table.t
   ; lib_config : Lib_config.t
   }
+
+let builtins db = db.builtins
 
 let paths t = t.paths
 
@@ -428,6 +432,8 @@ end = struct
         Dune_package.Entry.version e)
     ; entries
     ; dir
+    ; sections = Section.Map.empty
+    ; sites = Section.Site.Map.empty
     }
 
   let load_and_convert db ~dir ~meta_file ~name =