+++ src/dune_rules/findlib/findlib.mli
@@ -14,6 +14,9 @@ val create : paths:Path.t list -> lib_config:Lib_config.t -> t
 (** The search path for this DB *)
 val paths : t -> Path.t list
 
+(** The builtins packages *)
+val builtins : t -> Meta.Simplified.t Package.Name.Map.t
+
 val findlib_predicates_set_by_dune : Variant.Set.t
 
 module Unavailable_reason : sig