+++ src/dune_rules/super_context.ml
@@ -17,6 +17,8 @@ module Env_tree : sig
 
   val get_node : t -> dir:Path.Build.t -> Env_node.t
 
+  val get_context_env : t -> Env.t
+
   val create :
        context:Context.t
     -> host_env_tree:t option
@@ -25,6 +27,7 @@ module Env_tree : sig
     -> stanzas_per_dir:Stanza.t list Dir_with_dune.t Path.Build.Map.t
     -> root_expander:Expander.t
     -> bin_artifacts:Artifacts.Bin.t
+    -> context_env:Env.t
     -> t
 
   val bin_artifacts_host : t -> dir:Path.Build.t -> Artifacts.Bin.t
@@ -33,6 +36,7 @@ module Env_tree : sig
 end = struct
   type t =
     { context : Context.t
+    ; context_env : Env.t  (** context env with additional variables *)
     ; scopes : Scope.DB.t
     ; default_env : Env_node.t Memo.Lazy.t
     ; stanzas_per_dir : Stanza.t list Dir_with_dune.t Path.Build.Map.t
@@ -43,6 +47,8 @@ end = struct
     }
 
   let get_node t ~dir = t.get_node dir
+
+  let get_context_env t = t.context_env
 
   let bin_artifacts_host t ~dir =
     let bin_artifacts t ~dir = get_node t ~dir |> Env_node.bin_artifacts in
@@ -121,7 +127,7 @@ end = struct
     in
     Env_node.make ~dir ~scope ~config_stanza ~inherit_from:(Some inherit_from)
       ~profile:t.context.profile ~expander ~expander_for_artifacts
-      ~default_context_flags ~default_env:t.context.env
+      ~default_context_flags ~default_env:t.context_env
       ~default_bin_artifacts:t.bin_artifacts
 
   (* Here we jump through some hoops to construct [t] as well as create a
@@ -136,7 +142,7 @@ end = struct
      recursive module [Rec]. Since recursive let-modules are not allowed either,
      we need to also wrap [Rec] inside a non-recursive module [Non_rec]. *)
   let create ~context ~host_env_tree ~scopes ~default_env ~stanzas_per_dir
-      ~root_expander ~bin_artifacts =
+      ~root_expander ~bin_artifacts ~context_env =
     let module Non_rec = struct
       module rec Rec : sig
         val env_tree : unit -> t
@@ -145,6 +151,7 @@ end = struct
       end = struct
         let env_tree =
           { context
+          ; context_env
           ; scopes
           ; default_env
           ; stanzas_per_dir
@@ -199,6 +206,8 @@ type t =
 
 let context t = t.context
 
+let context_env t = Env_tree.get_context_env t.env_tree
+
 let stanzas t = t.stanzas
 
 let stanzas_in t ~dir = Path.Build.Map.find t.stanzas_per_dir dir
@@ -214,6 +223,27 @@ let to_dyn_concise t = Context.to_dyn_concise t.context
 let to_dyn t = Context.to_dyn t.context
 
 let host t = Option.value t.host ~default:t
+
+let get_site_of_packages t ~pkg ~site =
+  let find_site sites ~pkg ~site =
+    match Section.Site.Map.find sites site with
+    | Some section -> section
+    | None ->
+      User_error.raise
+        [ Pp.textf "Package %s doesn't define a site %s"
+            (Package.Name.to_string pkg)
+            (Section.Site.to_string site)
+        ]
+  in
+  match Package.Name.Map.find t.packages pkg with
+  | Some p -> find_site p.sites ~pkg ~site
+  | None -> (
+    match Findlib.find_root_package t.context.findlib pkg with
+    | Ok p -> find_site p.sites ~pkg ~site
+    | Error Not_found ->
+      User_error.raise
+        [ Pp.textf "The package %s is not found" (Package.Name.to_string pkg) ]
+    | Error (Invalid_dune_package exn) -> Exn.raise exn )
 
 let lib_entries_of_package t pkg_name =
   Package.Name.Map.find t.lib_entries_by_package pkg_name
@@ -393,10 +423,11 @@ let get_installed_binaries stanzas ~(context : Context.t) =
               acc)
       in
       match (stanza : Stanza.t) with
-      | Dune_file.Install { section = Bin; files; _ } ->
+      | Dune_file.Install { section = Section Bin; files; _ } ->
         binaries_from_install files
       | Dune_file.Executables
-          ({ install_conf = Some { section = Bin; files; _ }; _ } as exes) ->
+          ( { install_conf = Some { section = Section Bin; files; _ }; _ } as
+          exes ) ->
         let compile_info =
           let project = Scope.project d.scope in
           let dune_version = Dune_project.dune_version project in
@@ -458,6 +489,46 @@ let create ~(context : Context.t) ?host ~projects ~packages ~stanzas =
       ~context ~lib_artifacts:artifacts.public_libs
       ~bin_artifacts_host:artifacts_host.bin ~find_package
   in
+  let dune_dir_locations_var : Stdune.Env.Var.t = "DUNE_DIR_LOCATIONS" in
+  let env_dune_dir_locations =
+    let install_dir = Config.local_install_dir ~context:context.Context.name in
+    let install_dir = Path.build install_dir in
+    let v =
+      Option.value
+        (Stdune.Env.get context.env dune_dir_locations_var)
+        ~default:""
+    in
+    Package.Name.Map.foldi ~init:v packages ~f:(fun package_name package init ->
+        let sections =
+          Section.Site.Map.fold ~init:Install.Section.Set.empty
+            package.Package.sites ~f:(fun section acc ->
+              Install.Section.Set.add acc section)
+        in
+        let paths =
+          Install.Section.Paths.make ~package:package_name ~destdir:install_dir
+            ()
+        in
+        Install.Section.Set.fold sections ~init ~f:(fun section acc ->
+            sprintf "%s%c%s%c%s%s"
+              (Package.Name.to_string package_name)
+              Stdune.Bin.path_sep
+              (Section.to_string section)
+              Stdune.Bin.path_sep
+              (Path.to_absolute_filename
+                 (Install.Section.Paths.get paths section))
+              ( if String.is_empty acc then
+                acc
+              else
+                sprintf "%c%s" Stdune.Bin.path_sep acc )))
+  in
+  let context_env =
+    if String.is_empty env_dune_dir_locations then
+      context.env
+    else
+      Stdune.Env.add context.env ~var:dune_dir_locations_var
+        ~value:env_dune_dir_locations
+  in
+
   let default_env =
     Memo.lazy_ (fun () ->
         let make ~inherit_from ~config_stanza =
@@ -472,7 +543,7 @@ let create ~(context : Context.t) ?host ~projects ~packages ~stanzas =
           let expander = Memo.Lazy.of_val root_expander in
           Env_node.make ~dir ~scope ~inherit_from ~config_stanza
             ~profile:context.profile ~expander ~expander_for_artifacts
-            ~default_context_flags ~default_env:context.env
+            ~default_context_flags ~default_env:context_env
             ~default_bin_artifacts:artifacts.bin
         in
         make ~config_stanza:context.env_nodes.context
@@ -485,7 +556,7 @@ let create ~(context : Context.t) ?host ~projects ~packages ~stanzas =
   let env_tree =
     Env_tree.create ~context ~scopes ~default_env ~stanzas_per_dir
       ~host_env_tree:(Option.map host ~f:(fun x -> x.env_tree))
-      ~root_expander ~bin_artifacts:artifacts.bin
+      ~root_expander ~bin_artifacts:artifacts.bin ~context_env
   in
   let dir_status_db = Dir_status.DB.make ~stanzas_per_dir in
   let projects_by_key =