+++ test/blackbox-tests/test-cases/special_files.t/run.t
@@ -1,0 +1,5 @@
+  $ mkfifo src/foo
+  $ dune build @all
+  $ _build/default/bar.exe
+  hi!
+  $ rm -f src/foo